requirements.txt
vocode[io]==0.1.95
openai==0.27.2
python-dotenv==1.0.0